+++ packages/ccda/src/index.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 export * from './ccda-to-fhir';
 export * from './datetime';
-export * from './fhir-to-ccda';
+export * from './fhir-to-ccda/convert';
 export * from './systems';
 export * from './templates';
 export * from './types';
